% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibility.R
\name{bd_is_purgeable}
\alias{bd_is_purgeable}
\title{Can this plot library's output be dropped from the DOM and redrawn?}
\usage{
bd_is_purgeable(plotlib)
}
\description{
Can this plot library's output be dropped from the DOM and redrawn?
}
\keyword{internal}
