% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibility.R
\name{bd_purge_enabled}
\alias{bd_purge_enabled}
\title{Resolve a purge switch given as logical, function or reactive}
\usage{
bd_purge_enabled(purge)
}
\description{
Resolve a purge switch given as logical, function or reactive
}
\keyword{internal}
