% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module.R
\name{bd_record_download}
\alias{bd_record_download}
\title{Record a plot download}
\usage{
bd_record_download(ns)
}
\description{
No-op unless the host app registered a \code{bigdash.record_download} hook.
}
\keyword{internal}
