% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{bigdash.hideMenuElement}
\alias{bigdash.hideMenuElement}
\title{Hide a sidebar menu section by its label}
\usage{
bigdash.hideMenuElement(session, name)
}
\arguments{
\item{session}{Shiny session.}

\item{name}{Visible label of the section.}
}
\description{
Hide a sidebar menu section by its label
}
