% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{bigdash.hideMenuItem}
\alias{bigdash.hideMenuItem}
\title{Hide a sidebar menu item}
\usage{
bigdash.hideMenuItem(session, item)
}
\arguments{
\item{session}{Shiny session.}

\item{item}{\code{data-target} of the menu item.}
}
\description{
Hide a sidebar menu item
}
