% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{bigdash.hideTab}
\alias{bigdash.hideTab}
\title{Hide a tab and its menu item}
\usage{
bigdash.hideTab(session, tab)
}
\arguments{
\item{session}{Shiny session.}

\item{tab}{\code{data-name} of the tab.}
}
\description{
Hide a tab and its menu item
}
