% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{bigdash.removeTab}
\alias{bigdash.removeTab}
\title{Remove a tab, its settings pane and its menu item}
\usage{
bigdash.removeTab(session, tab)
}
\arguments{
\item{session}{Shiny session.}

\item{tab}{\code{data-name} of the tab.}
}
\description{
Remove a tab, its settings pane and its menu item
}
