% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{bigdash.selectTab}
\alias{bigdash.selectTab}
\title{Select a tab}
\usage{
bigdash.selectTab(session, selected)
}
\arguments{
\item{session}{Shiny session.}

\item{selected}{\code{data-target} of the tab to select.}
}
\description{
Select a tab
}
