% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{bigdash.showMenuElement}
\alias{bigdash.showMenuElement}
\title{Show a sidebar menu section by its label}
\usage{
bigdash.showMenuElement(session, name)
}
\arguments{
\item{session}{Shiny session.}

\item{name}{Visible label of the section.}
}
\description{
Show a sidebar menu section by its label
}
