% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{bigdash.showMenuItem}
\alias{bigdash.showMenuItem}
\title{Show a sidebar menu item}
\usage{
bigdash.showMenuItem(session, item)
}
\arguments{
\item{session}{Shiny session.}

\item{item}{\code{data-target} of the menu item.}
}
\description{
Show a sidebar menu item
}
