% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{bigdash.showTab}
\alias{bigdash.showTab}
\title{Show a tab and its menu item}
\usage{
bigdash.showTab(session, tab)
}
\arguments{
\item{session}{Shiny session.}

\item{tab}{\code{data-name} of the tab.}
}
\description{
Show a tab and its menu item
}
