% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{bigdash.showTabs}
\alias{bigdash.showTabs}
\title{Reveal the sidebar menu after the landing page}
\usage{
bigdash.showTabs(session)
}
\arguments{
\item{session}{Shiny session.}
}
\description{
Reveal the sidebar menu after the landing page
}
