% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{bigdash.toggleMenuItem}
\alias{bigdash.toggleMenuItem}
\title{Show or hide a sidebar menu item}
\usage{
bigdash.toggleMenuItem(session, item, state)
}
\arguments{
\item{session}{Shiny session.}

\item{item}{\code{data-target} of the menu item.}

\item{state}{\code{TRUE} to show, \code{FALSE} to hide.}
}
\description{
Show or hide a sidebar menu item
}
