% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{bigdash.toggleSidebar}
\alias{bigdash.toggleSidebar}
\title{Open or close the sidebar}
\usage{
bigdash.toggleSidebar(state)
}
\arguments{
\item{state}{\code{TRUE} to open, \code{FALSE} to close.}
}
\description{
Open or close the sidebar
}
