% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{bigdash.toggleTab}
\alias{bigdash.toggleTab}
\title{Show or hide a tab}
\usage{
bigdash.toggleTab(session, tab, state)
}
\arguments{
\item{session}{Shiny session.}

\item{tab}{\code{data-name} of the tab.}

\item{state}{\code{TRUE} to show, \code{FALSE} to hide.}
}
\description{
Show or hide a tab
}
