% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{bigdash.unloadSidebar}
\alias{bigdash.unloadSidebar}
\title{Collapse the sidebar back to its initial state}
\usage{
bigdash.unloadSidebar()
}
\description{
Collapse the sidebar back to its initial state
}
