% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BIGDASH_DEFAULT_ID}
\alias{BIGDASH_DEFAULT_ID}
\title{Default namespace id for \code{\link[=bigPage]{bigPage()}} and friends.}
\usage{
BIGDASH_DEFAULT_ID
}
\description{
Default namespace id for \code{\link[=bigPage]{bigPage()}} and friends.
}
\keyword{internal}
