% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{bigTabItem}
\alias{bigTabItem}
\title{Tab}
\usage{
bigTabItem(name, ...)
}
\arguments{
\item{name}{Name of tab.
This is to be reference in the \code{target} argument
of the \code{\link[=sidebarMenuElement]{sidebarMenuElement()}}.}

\item{...}{Content of the tab.}
}
\description{
Create a tab.
}
