% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{bigTabs}
\alias{bigTabs}
\title{Tabs}
\usage{
bigTabs(..., id = BIGDASH_DEFAULT_ID)
}
\arguments{
\item{...}{Tabs, must be \code{\link[=bigTabItem]{bigTabItem()}}.}

\item{id}{Namespace id, must match the \code{id} passed to the enclosing
\code{\link[=bigPage]{bigPage()}}. See \code{\link[=bigPage]{bigPage()}} for details on nesting.}
}
\description{
Create big tabs.
}
