% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-module.R
\name{color_from_middle}
\alias{color_from_middle}
\title{Diverging in-cell bar background for a DT column}
\usage{
color_from_middle(data, color1, color2)
}
\arguments{
\item{data}{The column values, used to scale the bars.}

\item{color1}{Colour for negative values.}

\item{color2}{Colour for positive values.}
}
\description{
Returns a \code{DT::JS()} gradient that draws bars left of centre for negative
values and right of centre for positive ones.
}
