% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{fullPage}
\alias{fullPage}
\title{Full page}
\usage{
fullPage(..., .class = "")
}
\arguments{
\item{...}{Content of the page.}

\item{.class}{Additional class to pass to parent
\verb{<div>}.}
}
\description{
Full page component, takes the entirety of the screen.
}
