% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hover-dropdown.R
\name{hover_dropdown}
\alias{hover_dropdown}
\title{Create a Hover Dropdown UI Element}
\usage{
hover_dropdown(..., icon_class = "fa fa-cog")
}
\arguments{
\item{...}{Arbitrary Shiny UI elements to display in the dropdown.}

\item{icon_class}{A character string that defines the class of the Font Awesome icon to use.}
}
\value{
A Shiny UI element that can be added to the UI definition of the app.
}
\description{
This function creates a hover dropdown UI element in a Shiny app. The dropdown appears
when the user hovers over the icon.
}
\examples{
hover_dropdown(
  icon_class = "fa fa-cog",
  div(class = "custom-item", "Item 1"),
  div(class = "custom-item", "Item 2")
)
}
