% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modalTrigger}
\alias{modalTrigger}
\title{Modal trigger}
\usage{
modalTrigger(id, target, ..., class = "")
}
\arguments{
\item{id}{Id of the trigger itself.}

\item{target}{Id of the modal to open.}

\item{...}{Content of the link, typically an icon.}

\item{class}{Extra classes for the link.}
}
\description{
A link that opens the modal with the given id. Pair with \code{\link[=modalUI]{modalUI()}}.
}
