% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modalUI}
\alias{modalUI}
\title{Modal}
\usage{
modalUI(
  id,
  title,
  ...,
  size = c("default", "sm", "lg", "xl", "fullscreen"),
  track_open = FALSE,
  footer = shiny::tags$div(class = "modal-footer", shiny::tags$button(type = "button",
    class = "btn btn-secondary", `data-bs-dismiss` = "modal", "Close"))
)
}
\arguments{
\item{id}{Modal id, referenced by the trigger's \code{target}.}

\item{title}{Modal title.}

\item{...}{Modal body.}

\item{size}{Modal size.}

\item{track_open}{Report the open/closed state to the server as the
\verb{<id>_is_open} input.}

\item{footer}{Modal footer; \code{NULL} for none.}
}
\description{
A Bootstrap 5 modal, declared inline in the UI and opened by a
\code{\link[=modalTrigger]{modalTrigger()}} rather than by \code{\link[shiny:showModal]{shiny::showModal()}}.
}
