% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{navbarDropdown}
\alias{navbarDropdown}
\title{Navbar Dropdown}
\usage{
navbarDropdown(title, ..., auto_close = "true")
}
\arguments{
\item{title}{Title of the dropdown, displayed on
the navbar.}

\item{...}{Content of the dropdown, see
\code{\link[=navbarDropdownTab]{navbarDropdownTab()}} and \code{\link[=navbarDropdownItem]{navbarDropdownItem()}}.}
}
\description{
A navbar dropdown menu.
}
