% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{navbarDropdownItem}
\alias{navbarDropdownItem}
\title{Dropdown Item}
\usage{
navbarDropdownItem(title, ..., link = "#")
}
\arguments{
\item{title}{Title of the button.}

\item{...}{Attributes passed to \verb{<a>} tag.}

\item{link}{Link the item opens.}
}
\description{
A dropdown item to use in \code{\link[=navbarDropdown]{navbarDropdown()}}:
}
