% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{navbarDropdownTab}
\alias{navbarDropdownTab}
\title{Navbar Dropdown Tab}
\usage{
navbarDropdownTab(title, target)
}
\arguments{
\item{title}{Title of the button.}

\item{target}{Target tab this should display.}
}
\description{
A navbar dropdown item that opens a tab,
used in \code{\link[=navbarDropdown]{navbarDropdown()}}.
}
