% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{navbarItem}
\alias{navbarItem}
\title{Navbar Item}
\usage{
navbarItem(..., .class = "")
}
\arguments{
\item{...}{Content of the navbar button.}

\item{.class}{Additional class to pass to \code{nav-item}}
}
\description{
A navbar item to use in \code{\link[=navbar]{navbar()}}.
}
