% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{navbarTab}
\alias{navbarTab}
\title{Navbar Tab}
\usage{
navbarTab(target, ...)
}
\arguments{
\item{target}{Target \code{\link[=bigTabItem]{bigTabItem()}} this should
show.}

\item{...}{Content of the navbar button.}
}
\description{
A navbar item to use in \code{\link[=navbar]{navbar()}} which
toggles a \code{\link[=bigTabItem]{bigTabItem()}}.
}
