% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{settings}
\alias{settings}
\title{Settings}
\usage{
settings(title = "Settings", ..., .posthook = NULL, id = BIGDASH_DEFAULT_ID)
}
\arguments{
\item{title}{Title of the sidebar.}

\item{...}{Content of the sidebar.}

\item{.posthook}{JavaScript function as string to run
after the settings have been rendered.}

\item{id}{Namespace id, must match the \code{id} passed to the enclosing
\code{\link[=bigPage]{bigPage()}}. See \code{\link[=bigPage]{bigPage()}} for details on nesting.}
}
\description{
Settings content to pass to \code{\link[=bigPage]{bigPage()}}.
}
