% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-help.R
\name{sidebarHelp}
\alias{sidebarHelp}
\alias{sidebarTabHelp}
\title{Sidebar Help}
\usage{
sidebarHelp(..., id = BIGDASH_DEFAULT_ID)

sidebarTabHelp(target, title, text)
}
\arguments{
\item{...}{Elements defined by \code{\link[=sidebarTabHelp]{sidebarTabHelp()}}.}

\item{id}{Namespace id, must match the \code{id} passed to the enclosing
\code{\link[=bigPage]{bigPage()}}. See \code{\link[=bigPage]{bigPage()}} for details on nesting.}

\item{target}{Target tab when this should be displayed.}

\item{title, text}{Title, and text of the help.}
}
\description{
Setup contextual help to display on each tab.
}
