% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebarItem}
\alias{sidebarItem}
\title{Sidebar Item}
\usage{
sidebarItem(title, target)
}
\arguments{
\item{title}{Title of the sidebar.}

\item{target}{Target \code{\link[=bigTabItem]{bigTabItem()}} this should toggle.}
}
\description{
A sidebar item to place as child of \code{\link[=sidebar]{sidebar()}}.
}
