% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{tabSettings}
\alias{tabSettings}
\title{Tab Settings}
\usage{
tabSettings(...)
}
\arguments{
\item{...}{Content.}
}
\description{
Tab settings, to place within \code{\link[=bigTabItem]{bigTabItem()}},
every \code{\link[=bigTabItem]{bigTabItem()}} expect max one \code{\link[=tabSettings]{tabSettings()}}.
These ultimately appear in the settings sidebar.
}
