% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{tabSettingsItem}
\alias{tabSettingsItem}
\title{Tab Settings Item}
\usage{
tabSettingsItem(element)
}
\arguments{
\item{element}{Content.}
}
\description{
Tab settings item to pass to \code{\link[=tabSettings]{tabSettings()}},
these are generally inputs that will appear in the
settings siderbar.
}
