% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-module.R
\name{trunc_display_row}
\alias{trunc_display_row}
\title{Truncate row to display. Must be placed as list in columnDefs list
under options of DT::datatable function.}
\usage{
trunc_display_row(target, length)
}
\arguments{
\item{target}{Target column}

\item{length}{Maximum string length}
}
\description{
Truncate row to display. Must be placed as list in columnDefs list
under options of DT::datatable function.
}
