% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-module.R
\name{ui.DataTable}
\alias{ui.DataTable}
\title{Standard/default datatable.}
\usage{
ui.DataTable(
  df,
  rownames = TRUE,
  num.cols = NULL,
  color.cols = NULL,
  substr.cols = NULL,
  substr.len = 60,
  ...
)
}
\arguments{
\item{df}{Data frame to render.}

\item{rownames}{Show rownames.}

\item{num.cols}{Columns to format to 4 significant digits.}

\item{color.cols}{Columns to shade with \code{\link[=color_from_middle]{color_from_middle()}}.}

\item{substr.cols}{Columns to truncate, with the full value in a tooltip.}

\item{substr.len}{Truncation length, recycled over \code{substr.cols}.}

\item{...}{Unused.}
}
\description{
Standard/default datatable.
}
