% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{bigdash.openSettings}
\alias{bigdash.openSettings}
\title{Open the settings drawer}
\usage{
bigdash.openSettings(lock = TRUE)
}
\arguments{
\item{lock}{Keep the drawer open when the pointer leaves it.}
}
\description{
Open the settings drawer
}
