% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{bigPage}
\alias{bigPage}
\title{Page}
\usage{
bigPage(
  ...,
  sidebar = htmltools::tagList(),
  settings = htmltools::tagList(),
  settings_position = c("right", "left"),
  title = "BigOmics",
  navbar = NULL,
  lang = NULL,
  theme = big_theme()
)
}
\arguments{
\item{...}{Content of the application.}

\item{sidebar}{Sidebar content as returned by \code{sidebar}.}

\item{settings}{Settings drawer.}

\item{settings_position}{Position of the settings bar.}

\item{title}{Title of the application.}

\item{navbar}{Navbar.}

\item{lang}{Language for meta tag.}

\item{theme}{Theme, passed to \code{theme}, defaults to
\code{\link[=big_theme]{big_theme()}} which returns an object of class
\code{bs_theme}. Can be changed but must be a n object
of the same class as returned by \code{\link[bslib:bs_theme]{bslib::bs_theme()}}
for \strong{Bootstrap 5.}}
}
\description{
Main application page.
}
