% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{bigPage}
\alias{bigPage}
\title{Page}
\usage{
bigPage(..., sidebar = htmltools::tagList(), title = "BigOmics", lang = NULL)
}
\arguments{
\item{...}{Content of the application.}

\item{sidebar}{Sidebar content as returned by \code{sidebar}.}

\item{title}{Title of the application.}

\item{lang}{Language for meta tag.}
}
\description{
Main application page.
}
