% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draggable.R
\name{masonry}
\alias{masonry}
\alias{draggable}
\alias{draggableItem}
\title{Masonry}
\usage{
draggable(inputId, ...)

draggableItem(inputId, ..., width = NULL, class = "")
}
\arguments{
\item{inputId}{Id of input.}

\item{...}{Content, HTML tags.}

\item{width}{Used to defined column width, e.g.: 6 results in col-6}

\item{class}{Any additional class to pass to parent div.}
}
\description{
Create a masonry grid.
}
