% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{navbar}
\alias{navbar}
\title{Navbar}
\usage{
navbar(title, ...)
}
\arguments{
\item{title}{Brand of the navbar.}

\item{...}{Content of the navbar, generally \verb{navbar*}
functions.}
}
\description{
Main navbar to pass to \code{\link[=bigPage]{bigPage()}} \code{navbar} argument.
}
