% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{settings}
\alias{settings}
\title{Settings}
\usage{
settings(title = "Settings", ...)
}
\arguments{
\item{title}{Title of the sidebar.}

\item{...}{Content of the sidebar.}
}
\description{
Settings content to pass to \code{\link[=bigPage]{bigPage()}}.
}
