% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebar}
\alias{sidebar}
\title{Sidebar}
\usage{
sidebar(title = "Menu", ..., top_expanded = NULL, top_collapsed = NULL)
}
\arguments{
\item{title}{Title of the sidebar.}

\item{...}{Content of the sidebar.}

\item{top_expanded, top_collapsed}{Top of page content, above title.}
}
\description{
Sidebar content to pass to \code{\link[=bigPage]{bigPage()}}.
}
