% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebar}
\alias{sidebar}
\title{Sidebar}
\usage{
sidebar(title = "Menu", ...)
}
\arguments{
\item{title}{Title of the sidebar.}

\item{...}{Content of the sidebar.}
}
\description{
Sidebar content to pass to \code{\link[=bigPage]{bigPage()}}.
}
