% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{sidebarMenu}
\alias{sidebarMenu}
\title{Sidebar Menu}
\usage{
sidebarMenu(text, icon, ...)
}
\arguments{
\item{text}{Text to display when the sidebar is expanded.}

\item{icon}{Icon to display when the sidebar is collapsed.}

\item{...}{Children, \code{\link[=sidebarMenuElement]{sidebarMenuElement()}}.}
}
\description{
Menu for the side bar to pass to \code{\link[=sidebar]{sidebar()}}.
}
