% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draggable.R
\name{swappable}
\alias{swappable}
\alias{swappableItem}
\alias{update_swappable}
\title{Swappable}
\usage{
swappable(inputId, ...)

swappableItem(inputId, ..., class = "")

update_swappable(id, order, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{...}{Content, HTML tags.}

\item{class}{Any additional class to pass to parent div.}

\item{id, inputId}{Id of input.}

\item{order}{List of \code{inputId}s of \code{draggableItem} in desired
order.}

\item{session}{A value reactive domain.}
}
\description{
Create swappable elements.
\code{swappableItem}s within \code{swappable} can be swapper.
}
